Delete variables in a batch in normalize_entity
Dropping the "additional_variables" from the original dataframe takes a
significant portion of the time, and this increases with the number of
"additional_variables".

Currently we call df.drop once for each variables. The number of
variables passed to df.drop does not affect the running time (in my
benchmarking), so dropping them all at once is O(1) instead of O(n).