Cache feature names
Computing the name of a feature can be expensive because primitives use
reflection to generate their names, and when features are stacked the
number of calls to get_name() grows. This commit caches the name when it
is computed the first time.

With max_depth=2 this sped up ft.dfs (just building the features) by
about 3x, and the improvement is greater for higher values of max_depth.