Cache feature names (#536)
Computing the name of a feature can be expensive because primitives use
reflection to generate their names, and when features are stacked the
number of calls to get_name() grows. This commit caches the name when it
is computed the first time.

On the flights dataset with max_depth=2 this sped up generating features
by about 3x. With more nesting the improvement will be greater.