Replace find_forward_path with find_forward_paths
And the same for the backward version.

The new methods are generators which yield each path. This allows us to
stop searching if we only need one path.

Note that we ignore paths with cycles, meaning that a path may be said
to be unique if the only other paths contain cycles.