Add relationship_path to agg and direct features (#544)
This adds an optional constructor parameter, and stores the
relationship_path on the feature. If no relationship_path is passed to
the constructor we search for a path between the target entity and base
entity. If none or more than one is found we raise an error.  When there
are more than one possible paths between entities the feature name will
include the full path.

This does not update any other logic to use the paths, as this will be
added in later commits.

- Add parent_name and child_name to Relationship.
- Update features JSON schema version and add logic for detecting old
  version.
- Replace find_forward_path with find_forward_paths. And the same for
  the backward version. The new methods are generators which yield
  each path. This allows us to stop searching if we only need one path.
  Note that we ignore paths with cycles, meaning that a path may be said
  to be unique if the only other paths contain cycles.
- Remove Mean.generate_name. There is no reason to override the
  parent method here.
- Update feature copy to use path. These were previously overlooked,
  and untested.