Store full entity dfs in separate trie
If there is an aggregation of a uses_full_entity feature the aggregation
should not be calculated on the full entity. Previously because we only
had one set of dfs this was not possible.

Now, we always store the "filtered" df in df_trie, and if necessary also
store a full df in large_df_trie.

If any features at a given node require a full df we first calculate
them. Then we extract the filtered df from the result and use this to
calculate any remaining features. Because FeatureSet.uses_full_entity is
True for features which have a dependent uses_full_entity feature the
large df does not need to contain any of the not_full_entity_features.