Change feature trie values to tuples
Of (needs full entity bool, set of need full features, set of rest of
features).

The main issue this solves is when there is a node without any features
we might still need the full entity. For example: if the only target
feature is "PERCENTILE(MEAN(customers.transactions))" there are no
features on "customers", but we need the full entity so that the every
row in "transactions" can get its relationship variable to the target
entity.