Store full entity dfs in separate trie (#598)
If there is an aggregation of a uses_full_entity feature the aggregation
should not be calculated on the full entity. Previously because we only
had one set of dfs this was not possible. Now, we always store the 
"filtered" df in df_trie, and if necessary also store a full df in large_df_trie.

The values of feature_trie are now a 3-tuple of
(needs full entity, needs full entity features, rest of features).

The main issue this solves is when there is a node without any features
we might still need the full entity. For example: if the only target
feature is "PERCENTILE(MEAN(customers.transactions))" there are no
features on "customers", but we need the full entity so that the every
row in "transactions" can get its relationship variable to the target
entity.

If any features at a given node require a full df we first calculate
them. Then we extract the filtered df from the result and use this to
calculate any remaining features.