Do not calculate dependencies of precalculated features
Currently if there is a precalculated feature "A(B)" (and no other
feature depends on B) we will still re-calculate B even though the
resulting feature vector will never be used.

This commit solves this by making FeatureSet aware of which features
will be precalculated, so that it can exclude them and their
dependencies from the feature trie.