Build transforms in two steps
First on identity and aggregations, then on direct features.

This allows features to be constructed which are direct features of
aggregations of transforms on the target.
E.g. "log: sessions.MEAN(log.ABSOLUTE(value))"

Previously this was not being built because the transform was not built
when the aggregation was being built.