Generate transform features of direct features (#623)
Transform features are now built twice: first on identity and aggregation
features, then on direct features. We do not create transforms of single
direct features, or when all inputs are direct features with the same
relationship path. This would be redundant as
`WEEKDAY(customers.signup_date)` is equivalent to
`customers.WEEKDAY(signup_date)`, which should already have been
calculated.